pyproject.toml
[tool.poetry]
name = "linkedIn-automation"
version = "0.1.0"
description = ""
authors = ["Urich Oliveira <oliveira.urich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.15.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
